requirements.txt
datasets>=2.14
faiss-cpu>=1.7
numpy>=1.24
pandas>=2.0
Pillow>=9.0
scikit-learn>=1.3
sentence-transformers>=2.2
torch>=2.1
transformers>=4.36
vllm>=0.4.0
